Making plots look much better
Using now the barplot function with added text to get the % for each barplot. Cex is nice at 1 even if it could be dinamically adjusted for the number of bars. But if I want to remove some of the bar that are close from 0, then a fixed value should be good enough.