Adding the damages to the probDens plot
Now the damages bar will affect the output of the ProbDens plot. Still, its not very interesting still it will give you the prob to do 1 damage on a a 3 damages weapons. But it's gonna be usefull when introducing feels not pain kind of feature. It could be done be concatenating probabilities together to just be multiples of the damage characteristic of the weapon.